refactor IstexSummaryVIew to use compose renderPros, removing IstexYear and one file per component
