+++ src/app/js/formats/istexSummary/IstexSummaryView.js
@@ -1,19 +1,35 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import translate from 'redux-polyglot/translate';
+import compose from 'recompose/compose';
 
-import { field as fieldPropTypes } from '../../propTypes';
+import {
+    field as fieldPropTypes,
+    polyglot as polyglotPropTypes,
+} from '../../propTypes';
 import injectData from '../injectData';
-import IstexYear from './IstexYear';
 import InvalidFormat from '../InvalidFormat';
 import { getYearUrl, parseYearData } from './getIstexData';
 import { searchedFieldValues } from './IstexSummaryAdmin';
+import composeRenderProps from '../../lib/composeRenderProps';
 import IstexList from './IstexList';
+import IssueFold from './IssueFold';
+import VolumeFold from './VolumeFold';
+import YearFold from './YearFold';
+import IstexItem from '../istex/IstexItem';
+
+const FoldList = props => <IstexList {...props} />;
+const IstexDocument = ({ item }) => <IstexItem {...item} />;
+IstexDocument.propTypes = {
+    item: PropTypes.shape({ id: PropTypes.string.isRequired }).isRequired,
+};
 
 export const IstexSummaryView = ({
     formatData,
     field,
     resource,
     searchedField,
+    p: polyglot,
 }) => {
     if (!resource[field.name] || !searchedField) {
         return (
@@ -21,17 +37,22 @@ export const IstexSummaryView = ({
         );
     }
 
-    return (
-        <IstexList data={parseYearData(formatData)}>
-            {({ item: { name, count } }) => (
-                <IstexYear
-                    issn={resource[field.name]}
-                    year={name}
-                    count={count}
-                    searchedField={searchedField}
-                />
-            )}
-        </IstexList>
+    return composeRenderProps(
+        <FoldList
+            data={parseYearData(formatData)}
+            issn={resource[field.name]}
+            searchedField={searchedField}
+            polyglot={polyglot}
+        />,
+        [
+            YearFold,
+            FoldList,
+            VolumeFold,
+            FoldList,
+            IssueFold,
+            FoldList,
+            IstexDocument,
+        ],
     );
 };
 
@@ -42,6 +63,7 @@ IstexSummaryView.propTypes = {
     formatData: PropTypes.shape({}),
     error: PropTypes.string,
     searchedField: PropTypes.oneOf(searchedFieldValues),
+    p: polyglotPropTypes.isRequired,
 };
 
 IstexSummaryView.defaultProps = {
@@ -52,4 +74,4 @@ IstexSummaryView.defaultProps = {
     error: null,
 };
 
-export default injectData(getYearUrl)(IstexSummaryView);
+export default compose(injectData(getYearUrl), translate)(IstexSummaryView);
