+++ src/app/js/lib/composeRenderProps.js
@@ -1,0 +1,12 @@
+import { cloneElement } from 'react';
+
+export default (parent, renderProps) => {
+    const composedRender = renderProps.reverse().reduce((acc, render) => {
+        if (!acc) {
+            return render;
+        }
+        return (...args) => cloneElement(render(...args), { children: acc });
+    }, null);
+
+    return cloneElement(parent, {}, composedRender);
+};
